requirements.txt
# PDF处理库 - 使用预编译版本
PyMuPDF>=1.23.0

# 图像处理库
Pillow>=10.0.0

# 基于属性的测试库
hypothesis>=6.90.0

# 测试框架
pytest>=7.4.0

# 类型检查
mypy>=1.7.0

# 代码格式化
black>=23.0.0

# 打包工具
PyInstaller>=6.0.0